Comment in a reason about what happens when the user has no guid
The users.csv file is created through an offline process which removes
projects and users based on more information than just the
`application_state != edit` filter we use in this code. Therefore, it is
fine if the user id associated with a project can't be found in the
users.csv resource, it just means that we should ignore the project.