Correctly identify the call for each project
The "Call_Submitted" column in the synth source databases isn't a foreign key to
the "NHM_Call" table, it's just the numerical call order that the project was
submitted against. Therefore, we need to not lookup the "Call"'s id in the target
database using this value directly but choose the call in the analysis database
based on the order using the "Call_Submitted" column.