Add step to fill in as many empty home_institution_country fields as possible in VisitorProject
Using the cities we can generally figure out which country to put in this field if it's blank.
This approach uses a mix of looking up names in a Python lib called geonamescache which provides
lists of countries and cities and then having a static map of a few of the more tricky ones
(unmatched_home_institutions.json).

Where multiple countries were matched to a city and the city wasn't included in the unmatched json
file, the one with the highest population is used. This may seem arbitrary but it worked as a
differentiator for all but the rows with the city set to Islamabad. An example of where this works
is "Moscow" which should be Russia not the USA, or "Washington" which should be the USA not GB. In
the future this technique may need to be tweaked as synth round 4 is completed and more entries
are created which may break this currently working algorithm.